fix funny problem due to limitation(?) of R sub-assignment:
if f(a) returns an object of class X, for which we've defined `$<-.X`,
f(a)$b<-c doesn't work, which is to be expected, given subassignment
semantics, but(f(a))$b<-c also doesn't work, since it tries to find
a function `(<-`, which hasn't been defined.  Is there no way
to delimit the scope of subassignment?

Of course, x=f(a); x$b<-c works.

Closes #121