put all job folders on the NAS;
inotifywait works on this as long as the events are generated by this
server.  The only events which aren't would be file uploads, and these
will be handled by an API call instead.

Otherwise, we're trying to move job folders back and forth between
the local HD and NAS, which doesn't work from R's `file.rename()`