deal with weird forced-clock-jump NFS lock recovery failure
this daily event, likely related to the linux-under-MS-hypervisor
set-up, kills the connection to the NAS.  This isn't noticed by
mysqld until the innoDB storage engine has to touch backing files
(which are on NAS).

The existing "select 1" query didn't exercise this possiblity, so
we now do an update, which will.