check for empty vector of files to trash.
Saves calling dir(dst) etc. for no reason, and that call was
failing with this unexplained error:

Error: Error in dir(dst): 'translateCharUTF8' must be called on a CHARSXP

which is odd, because it's just a call to list a folder.  That folder does
have ~1E6 files, however, because I haven't been culling it or
archiving it.