add 6 new tag models
We assume these tags use batteries whose lifespan data are
already present in the file:

   inst/doc/lotekTagLifespanByBatteryAndBI.csv

But since there are no data for a "5-1" battery model, the tag
"NTQB2-5-1" is *not* included in the added tags.

The only user-made changes are in inst/doc/modelLotekTagLifeSpan.Rmd,
those in the other two versions of that file are generated by code.

To propagate these changes into the running servers, the following steps
were taken (as user sg):

   ## re-run modelling code and regenerate modelLotekTagLifeSpan.rda
   ## Due to no new battery-life data, this is just associating new models
   ## with existing batteries, but if there were new data in
   ## inst/doc/lotekTagLifespanByBatteryAndBI.csv the same procedure would
   ## be used.
   cd ~/src/motusServer/inst/doc
   make
   cd ../..
   ## rebuild the R motusServer package
   ~sg/bin/rpack -g .
   ## kill processServers gracefully (letting them finish their current
   ## subjobs first)
   /sgm/bin/killMotusProcessServers.sh -a -g
   ## wait a few seconds to avoid process creation / death race conditions
   sleep 5
   ## restart motus servers (those previously left running are unaffected)
   /sgm/bin/runAllMotusServers.sh